+++ src/lineage/__init__.py
@@ -22,11 +22,9 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 """
 
-import datetime
 from itertools import chain
 import os
 
-from atomicwrites import atomic_write
 import numpy as np
 import pandas as pd
 
@@ -35,7 +33,7 @@ from lineage.ensembl import EnsemblRestClient
 from lineage.individual import Individual
 from lineage.resources import Resources
 from lineage.snps import SNPs
-from lineage.utils import Parallelizer
+from lineage.utils import Parallelizer, create_dir, save_df_as_csv
 from lineage.visualization import plot_chromosomes
 
 # set version string with Versioneer
@@ -913,85 +911,3 @@ class Lineage:
                 continue
 
             self.remap_snps(i, 37)
-
-
-def create_dir(path):
-    """ Create directory specified by `path` if it doesn't already exist.
-
-    Parameters
-    ----------
-    path : str
-        path to directory
-
-    Returns
-    -------
-    bool
-        True if `path` exists
-    """
-    # https://stackoverflow.com/a/5032238
-    try:
-        os.makedirs(path, exist_ok=True)
-    except Exception as err:
-        print(err)
-        return False
-
-    if os.path.exists(path):
-        return True
-    else:
-        return False
-
-
-def save_df_as_csv(df, path, filename, comment=None, **kwargs):
-    """ Save dataframe to a CSV file.
-
-    Parameters
-    ----------
-    df : pandas.DataFrame
-        dataframe to save
-    path : str
-        path to directory where to save CSV file
-    filename : str
-        filename of CSV file
-    comment : str
-        header comment(s); one or more lines starting with '#'
-    **kwargs
-        additional parameters to `pandas.DataFrame.to_csv`
-
-    Returns
-    -------
-    str
-        path to saved file, else empty str
-    """
-    if isinstance(df, pd.DataFrame) and len(df) > 0:
-        try:
-            if not create_dir(path):
-                return ""
-
-            destination = os.path.join(path, filename)
-
-            print("Saving " + os.path.relpath(destination))
-
-            s = (
-                "# Generated by lineage v{}, https://github.com/apriha/lineage\n"
-                "# Generated at {} UTC\n"
-            )
-
-            s = s.format(
-                __version__, datetime.datetime.utcnow().strftime("%Y-%m-%d %H:%M:%S")
-            )
-
-            if isinstance(comment, str):
-                s += comment
-
-            with atomic_write(destination, mode="w", overwrite=True) as f:
-                f.write(s)
-                # https://stackoverflow.com/a/29233924
-                df.to_csv(f, na_rep="--", **kwargs)
-
-            return destination
-        except Exception as err:
-            print(err)
-            return ""
-    else:
-        print("no data to save...")
-        return ""
