+++ src/lineage/utils.py
@@ -1,5 +1,11 @@
+import datetime
 from multiprocessing import Pool
 import os
+
+from atomicwrites import atomic_write
+import pandas as pd
+
+import lineage
 
 
 class Parallelizer:
@@ -36,3 +42,86 @@ class Parallelizer:
                 return p.map(f, tasks)
         else:
             return map(f, tasks)
+
+
+def create_dir(path):
+    """ Create directory specified by `path` if it doesn't already exist.
+
+    Parameters
+    ----------
+    path : str
+        path to directory
+
+    Returns
+    -------
+    bool
+        True if `path` exists
+    """
+    # https://stackoverflow.com/a/5032238
+    try:
+        os.makedirs(path, exist_ok=True)
+    except Exception as err:
+        print(err)
+        return False
+
+    if os.path.exists(path):
+        return True
+    else:
+        return False
+
+
+def save_df_as_csv(df, path, filename, comment=None, **kwargs):
+    """ Save dataframe to a CSV file.
+
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        dataframe to save
+    path : str
+        path to directory where to save CSV file
+    filename : str
+        filename of CSV file
+    comment : str
+        header comment(s); one or more lines starting with '#'
+    **kwargs
+        additional parameters to `pandas.DataFrame.to_csv`
+
+    Returns
+    -------
+    str
+        path to saved file, else empty str
+    """
+    if isinstance(df, pd.DataFrame) and len(df) > 0:
+        try:
+            if not create_dir(path):
+                return ""
+
+            destination = os.path.join(path, filename)
+
+            print("Saving " + os.path.relpath(destination))
+
+            s = (
+                "# Generated by lineage v{}, https://github.com/apriha/lineage\n"
+                "# Generated at {} UTC\n"
+            )
+
+            s = s.format(
+                lineage.__version__,
+                datetime.datetime.utcnow().strftime("%Y-%m-%d %H:%M:%S"),
+            )
+
+            if isinstance(comment, str):
+                s += comment
+
+            with atomic_write(destination, mode="w", overwrite=True) as f:
+                f.write(s)
+                # https://stackoverflow.com/a/29233924
+                df.to_csv(f, na_rep="--", **kwargs)
+
+            return destination
+        except Exception as err:
+            print(err)
+            return ""
+    else:
+        print("no data to save...")
+        return ""
